add index html rendering
