+++ labs/unwrapped/Cargo.toml
@@ -12,3 +12,5 @@ tokio = { version = "1", features = ["full"] }
 tower-http = { version = "0.5.0", features = ["trace"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+rust-embed = { version = "8.4.0", features = ["include-exclude"] }
+tokio-util = { version = "0.6", features = ["io"] }
