+++ labs/unwrapped/assets/index.html
@@ -1,0 +1,50 @@
+<html>
+  <head>
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Outfit:wght@100..900&display=swap"
+      rel="stylesheet"
+    />
+    <style type="text/css">
+      html {
+        font-size: 20px;
+      }
+      body {
+        font-family: "Outfit", sans-serif;
+        font-optical-sizing: auto;
+        font-weight: 500;
+        font-style: normal;
+      }
+      main {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        text-align: center;
+        width: 100vw;
+        height: 100vh;
+      }
+      main h1 {
+        font-size: 10rem;
+        margin: 0 0.5rem;
+      }
+      main h2 {
+        color: gray;
+      }
+      main a {
+        color: #29b158;
+        text-decoration: none;
+      }
+    </style>
+  </head>
+  <body>
+    <main>
+      <h1>unwrapped</h1>
+      <h2>
+        a music collection software by
+        <a href="https://sneakycrow.dev">sneaky crow</a>
+      </h2>
+    </main>
+  </body>
+</html>
