+++ labs/unwrapped/src/assets.rs
@@ -1,0 +1,6 @@
+use rust_embed::Embed;
+
+#[derive(Embed)]
+#[folder = "assets/"]
+#[include = "*.html"]
+pub(crate) struct Assets;
