+++ labs/unwrapped/src/main.rs
@@ -1,3 +1,4 @@
+mod assets;
 mod db;
 mod routes;
